made FAQ and support communication

+++ handlers/message_editor.py
@@ -11,6 +11,7 @@ class TravelEditor:
     async def edit_message(lexicon_key: str, request, delete_mode=False, media_markup_mode=False,
                            button_texts: Set[str] = None, callback_sign: str = None):
         '''Высылальщик сообщений, '''
+        new_message = None
         lexicon_part = LEXICON[lexicon_key]
         redis_key = str(request.from_user.id) + ':last_message'
         last_message_id = await redis_data.get_data(redis_key)
@@ -29,7 +30,20 @@ class TravelEditor:
 
         message_text = lexicon_part['message_text']
 
-        if delete_mode:
+        user_id = request.from_user.id
+        redis_key_current_lexicon = str(user_id) + ':current_lexicon_code'
+        current_lexicon = await redis_data.get_data(key=redis_key_current_lexicon)
+        if current_lexicon:
+            redis_key_last_lexicon = str(user_id) + ':last_lexicon_code'
+            await redis_data.set_data(redis_key_last_lexicon, current_lexicon)
+            await redis_data.set_data(redis_key_current_lexicon, 0)
+
+        redis_key = str(user_id) + ':current_lexicon_code'
+        await redis_data.set_data(redis_key, lexicon_key)
+
+        redis_key = str(request.from_user.id) + ':last_message'
+
+        if delete_mode and last_message_id:
             await chat.Chat.delete_message(self=chat_object, message_id=last_message_id)
             new_message = await message_object.answer(text=message_text, reply_markup=keyboard)
             await redis_data.set_data(redis_key, new_message.message_id)
@@ -40,9 +54,15 @@ class TravelEditor:
             try:
                 await message_object.edit_text(text=message_text, reply_markup=keyboard)
             except:
-                await chat.Chat.delete_message(self=chat_object, message_id=last_message_id)
-                new_message = await message_object.answer(text=message_text, reply_markup=keyboard)
-                await redis_data.set_data(redis_key, new_message.message_id)
+                try:
+                    await chat.Chat.delete_message(self=chat_object, message_id=last_message_id)
+                except:
+                    pass
+                finally:
+                    new_message = await message_object.answer(text=message_text, reply_markup=keyboard)
+                    await redis_data.set_data(redis_key, new_message.message_id)
 
+        if new_message:
+            print('message = ', new_message.message_id)
 
 travel_editor = TravelEditor()
